### 2024-03-11 #6463 @editorialbot
[PRE REVIEW]: NemesisPy: A Python package for simulating and retrieving exoplanetary spectra
**Submitting author:** @Jingxuan97 (<a href="http://orcid.org/0009-0006-2395-6197">Jingxuan Yang</a>)
**Repository:** https://github.com/Jingxuan97/nemesispy
**Branch with paper.md** (empty if default branch):
**Version:** 0.0.10
