### 2024-03-11 #6463 @dfm
In the meantime, please take a look through the [JOSS review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html) to make sure that the code base fully satisfies all the criteria.
