### 2024-03-11 #6463 @Jingxuan97
Hi,

Here are my reviewer suggestions:

Lorenzo V. Mugnai, Sarah Blunt, Enrico Olivier, Shubh Agrawal.

Best regards,
Jingxuan

On Mon, Mar 11, 2024 at 12:34 AM Dan Foreman-Mackey <
***@***.***> wrote:

> @Jingxuan97 <https://github.com/Jingxuan97> — Thanks for your submission!
> All the suitable JOSS editors are currently working at capacity so I'm
> going to "waitlist" this review until an editor with the relevant expertise
> is available to take it on. Thanks for your patience!
>
> In the meantime, please take a look through the JOSS review criteria
> <https://joss.readthedocs.io/en/latest/review_criteria.html> to make sure
> that the code base fully satisfies all the criteria.
>
