### 2024-06-03 #6463 @ben-cassese
Sorry for the delay, I'd be happy to! Following @dfm's comment above though, it'd be helpful to clarify— is the documentation entirely contained within the [nemesispy.pdf](https://github.com/Jingxuan97/nemesispy/blob/joss/nemesispy.pdf) file? Also, could @Jingxuan97 and clarify the relationship between this code, nemesispy, and this code, [NemesisPy](https://nemesispy.readthedocs.io/en/latest/), which seems to also be under active development by a coauthor and has its own documentation/examples?
