### 2024-06-07 #6463 @Jingxuan97
Hi,

Regarding the NemesisPy repository, we have now made it private, so it
should not be visible online anymore. We plan to release that code base in
the future separately under a different name. We will be careful to
distinguish these two code bases going forward.

Regarding the documentation, it is currently contained entirely within the
nemesispy.pdf
<https://github.com/Jingxuan97/nemesispy/blob/joss/nemesispy.pdf> file. Is
the issue with the documentation itself, or how it is hosted? I.e., should
we host it online?

Best regards,
Jingxuan
