### 2024-06-11 #6874 @editorialbot
[REVIEW]: NemesisPy: A Python package for simulating and retrieving exoplanetary spectra
**Submitting author:** @Jingxuan97 (<a href="http://orcid.org/0009-0006-2395-6197">Jingxuan Yang</a>)
**Repository:** https://github.com/Jingxuan97/nemesispy
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.2
