### 2024-06-21 Jingxuan97/nemesispy#1 @ben-cassese
JOSS Review: Missing DOIs in paper
Here are the DOIs that the EditorialBot flagged as missing:

missing doi #1, for Barstow+ 2017: 10.3847/1538-4357/834/1/50
missing doi #2, for Yang+ 2024: 10.1093/mnras/stae1427

These need to be included in paper.bib.
