### 2024-06-21 Jingxuan97/nemesispy#2 @ben-cassese
JOSS Review: Pip installation fails
This is an easy fix: right now pip installing from source fails because `setup.py` looks for `README.md`
for the `long_description`, but at some point the README was changed to a `.rst` file.
