### 2024-06-21 Jingxuan97/nemesispy#3 @ben-cassese
JOSS Review: No community guidelines
This is a minor one, but currently this requirement for a JOSS paper is missing:

"Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support"
