### 2024-06-21 #6874 @ben-cassese
I've completed my initial review. Excellent package! This is a very useful tool for the community, thanks for open-sourcing your hard work. I've opened a handful of minor issues on the repository itself that shouldn't take long to fix. I've also opened one issue that is my only major blocker right now, which concerns the documentation of the package. More details are included [here](https://github.com/Jingxuan97/nemesispy/issues/4). Once the documentation is augmented/this issue is resolved, I'd be happy to recommend acceptance. Great work again!
