### 2024-06-21 #6874 @ben-cassese
For breadcrumbs, the minor issues are [here](https://github.com/Jingxuan97/nemesispy/issues/1), [here](https://github.com/Jingxuan97/nemesispy/issues/2), and [here](https://github.com/Jingxuan97/nemesispy/issues/3).
