### 2024-06-25 jingxuan97/nemesispy2022#1 @Jingxuan97
Missing DOIs now added
