### 2024-06-25 jingxuan97/nemesispy2022#2 @Jingxuan97
`setup.py` file changed to look for `README.rst`
