### 2024-06-26 jingxuan97/nemesispy2022#3 @Jingxuan97
Guidelines now added to the new documentation
