### 2024-06-26 jingxuan97/nemesispy2022#4 @Jingxuan97
Thank you for the suggestions, they are very helpful indeed. I have now done the following:

- Deployed the documentation to GitHub pages at https://jingxuan97.github.io/nemesispy/
- Added documentation for the ForwardModel class, both in the code and in the external documentation
- Added a tutorial going through the example calculation involving the ForwardModel class
- Included more functions in the API
