### 2024-07-08 jingxuan97/nemesispy2022#3 @shubhagrawal30
Adding onto this: it took me a decent amount of time to find these, including having to come to this issue to realize where they were added. It is probably best to also add a section to the GitHub repository README.
