### 2024-07-08 Jingxuan97/nemesispy#5 @shubhagrawal30
JOSS Review: State of the Field
minor comment: A brief description/list of other packages that can perform radiative transfer calculations and/or atmospheric modeling will be useful. Currently, you generally describe the usefulness of such packages and list special features of NEMESISpy, but there is no description of a collection of software that is also used for such an application.
