### 2024-07-08 Jingxuan97/nemesispy#6 @shubhagrawal30
JOSS Review: Functionality documentation
minor comment: the project README can be formatted better: putting installation instructions, testing, and demo code in their own sections; utilizing `the code markdown feature` when talking about lines of code: for eg., using `pip install --editable .`.

second minor comment: for the demo at [https://jingxuan97.github.io/nemesispy/tutorial.html](https://jingxuan97.github.io/nemesispy/tutorial.html), it would be useful to also have the outputs obtained from running each segment. So that a new user can compare their code against a standard / can take a look at the functionality of the package without needing to set it up locally.

third minor comment: could be useful to add in some comments about what each of the data files in `nemesispy/data` represent / were obtained from.
