### 2024-07-08 #6874 @shubhagrawal30
Only Minor Issues:
1. Added a note about _community guidelines_ to the [previously added issue#3](https://github.com/Jingxuan97/nemesispy/issues/3).
2. _State of the field:_ [issue#5](https://github.com/Jingxuan97/nemesispy/issues/5)
3. _Functionality documentation_: [issue#6](https://github.com/Jingxuan97/nemesispy/issues/6) (x3)
4.

Some notes:
1. _ Performance_: paper talks about just-in-time compilation to improve speeds for Bayesian retrieval ---> extensive use of numba jit, so looks good.
2. previous issues from other reviewer seem to have been taken care of.

P.S. @JBorrow: Generally speaking, I cannot claim to be a good judge of "Does the full list of paper authors seem appropriate and complete?" in _Contribution and authorship_. Not sure of the expectations, but as an independent reviewer, I have no clue who has worked on this project. I have checked the box, but defer to others including editorial staff.
