+++ nemesispy/radtran/scatter.py
@@ -1,0 +1,1316 @@
+import os
+import numpy as np
+from numba import njit,prange
+
+MAXMU=21
+DNUS0 = 354.69
+DNUS1 = 587.07
+DNUQ = 4161.0
+
+jcoef = np.array([
+    [1, 0, 0, 0, 0],
+    [1, 0, 0, 0, 0],
+    [-1, 3, 0, 0, 0],
+    [-3, 5, 0, 0, 0],
+    [3, -30, 35, 0, 0],
+    [15, -70, 63, 0, 0],
+    [-5, 105, -315, 231, 0],
+    [-35, 315, -693, 429, 0],
+    [35, -1260, 6930, -12012, 6435],
+    [315, -4620, 18018, -25740, 12155]
+], dtype=int)
+jdiv = np.array([1, 2, 8, 16, 128], dtype=int)
+
+@njit(fastmath=True, error_model='numpy')
+def miescat(xlam, iscat, dsize, nmode, rs, refindx, theta, ntheta, nphas):
+    pi = np.pi
+    maxth = 100
+    maxmode = 10
+    pi2 = 1.0 / np.sqrt(2.0 * pi)
+    func = np.zeros((4, 2 * maxth))
+    phas0 = np.zeros(2 * maxth)
+    phas = np.zeros(nphas)
+    idump = 0
+    numflag = False
+    
+    if ntheta > maxth:
+        print('  TOO MANY ANGLE POINTS: NTHETA =', ntheta, ' MAXTH =', maxth)
+    if nmode > maxmode:
+        print('  TOO MANY PARTICLE MODES; NMODE =', nmode, ' MAXMODE =', maxmode)
+    thetd = np.zeros_like(theta)
+    for j in range(ntheta):
+        if theta[j] < 0 or theta[j] > 90:
+            print(' ANGLE <0 OR > 90')
+            return
+        thetd[j] = theta[j]
+    r1 = rs[0]
+    delr = rs[2]
+    if rs[1] < rs[0]:
+        inr = 1000000001
+        cont0 = False
+    else:
+        inr = 1 + int((rs[1] - rs[0]) / rs[2])
+        if inr > 1 and inr % 2 != 0:
+            inr += 1
+        cont0 = True
+    nqmax = np.zeros(nmode)
+    if not cont0:
+        rmax = np.zeros(nmode)
+        for m in range(nmode):
+            nqmax[m] = 0.0
+            if dsize[m][1] != 0:
+                aa = dsize[m][0]
+                bb = dsize[m][1]
+                alpha = 0.0
+                cc = 0.0
+                if iscat[0] == 0: # continuous
+                    rmax[0] = rs[1]
+
+                if iscat[m] == 1:
+                    alpha = dsize[m][2]
+                    rmax[m] = alpha * aa * bb
+                elif iscat[m] == 2:
+                    rmax[m] = np.exp(np.log(aa) - bb ** 2)
+                else:
+                    cc = dsize[m][2]
+                    rmax[m] = (aa / (bb * cc)) ** (1.0 / cc)
+    if iscat[0] == 0: # continuous
+        r_grid = np.exp(np.linspace(np.log(r1),np.log(rs[1]),dsize.shape[1]))
+        r_dist_grid = dsize[0]
+        r_dist = np.interp(np.arange(r1,rs[1]+delr,delr),r_grid,r_dist_grid)
+    kscat = 0.0
+    area = 0.0
+    volume = 0.0
+    kext = 0.0
+    anorm = 0.0
+    rfr = refindx[0]
+    rfi = refindx[1]
+    for m in range(inr):
+#     while True:
+        rr = r1 + m * delr
+        csratio = -1.0
[…]
