### 2024-07-09 jingxuan97/nemesispy2022#5 @Jingxuan97
We have now added a 'State of the field' section with the following sentence "For a review of exoplanet atmospheric retrieval codes with comparable functionalities to NEMESISPY, we refer the reader to the comprehensive catalogue in @macdonald_catalog_2023.", which points the reader to a paper that contains a list of retrieval softwares that can perform varieties of radiative transfer and atmospheric modelling (https://iopscience.iop.org/article/10.3847/2515-5172/acc46a).
