### 2024-07-09 #6874 @Jingxuan97
Thank you @shubhagrawal30 <https://github.com/shubhagrawal30> for your
comments!

On Mon, Jul 8, 2024 at 3:19 PM Shubh Agrawal ***@***.***>
wrote:

> Only Minor Issues:
>
>    1. Added a note about *community guidelines* to the previously added
>    issue#3 <https://github.com/Jingxuan97/nemesispy/issues/3>.
>    2. *State of the field:* issue#5
>    <https://github.com/Jingxuan97/nemesispy/issues/5>
>    3. *Functionality documentation*: issue#6
>    <https://github.com/Jingxuan97/nemesispy/issues/6> (x3)
>    4.
>
> Some notes:
>
>    1. _ Performance_: paper talks about just-in-time compilation to
>    improve speeds for Bayesian retrieval ---> extensive use of numba jit, so
>    looks good.
>    2. previous issues from other reviewer seem to have been taken care of.
>
> P.S. @JBorrow <https://github.com/JBorrow>: Generally speaking, I cannot
> claim to be a good judge of "Does the full list of paper authors seem
> appropriate and complete?" in *Contribution and authorship*. Not sure of
> the expectations, but as an independent reviewer, I have no clue who has
> worked on this project. I have checked the box, but defer to others
> including editorial staff.
>
