### 2024-07-09 jingxuan97/nemesispy2022#6 @Jingxuan97
Minor comment #1: we have now reformatted the README file with the code markdown feature.
