### 2024-07-09 jingxuan97/nemesispy2022#3 @Jingxuan97
We have now added the guidelines to the README file.
