### 2024-07-09 jingxuan97/nemesispy2022#6 @Jingxuan97
Minor comment #2: we have added the output of running the emission spectrum calculation to the online tutorial.
