### 2024-07-09 jingxuan97/nemesispy2022#6 @Jingxuan97
Minor comment #3: we have now commented about the data files in `nemesispy/data` in the online documentation under `Usage/Data`, in particular where they can be obtained from.
