### 2024-07-10 jingxuan97/nemesispy2022#6 @shubhagrawal30
This looks good! One note, I think it would be very useful to also add the figure `emission_example.pdf` to the online tutorial.
