### 2024-07-10 #6874 @shubhagrawal30
Very minor additional note on https://github.com/Jingxuan97/nemesispy/issues/6, everything else looks good to go!
