### 2024-07-15 #6874 @JBorrow
Great! Thank you both for your insightful reviews.

@shubhagrawal30, regarding your question - on smaller projects like this that's less of a concern. When we have, for instance, 20 or so authors, we like to make sure that they all actually contributed to the project.

@shubhagrawal30 please let me know if there is anything outstanding on your end or not.
