+++ nemesispy/common/utils.py
@@ -51,21 +51,25 @@ def divide_gcm_grid(nlon,nlat,nrank):
     return partition
 
 def bin_down(input_value, input_grid, bin_edges):
-    """_summary_
+    """Bin down the input_value based on the input_grid and bin_edges.
 
     Args:
-        input_value (_type_): _description_
-        input_grid (_type_): _description_
-        bin_edges (_type_): _description_
+        input_value : ndarray
+            An array of values to be binned down.
+        input_grid : ndarray
+            The grid on which the input_value is defined.
+        bin_edges : ndarray
+            Edges of the bins to be used for binning down.
 
     Returns:
-        _type_: _description_
+        output_value : ndarray
+            An array of binned down values.
     """
-    out = np.zeros(len(bin_edges)-1)
+    output_value = np.zeros(len(bin_edges)-1)
     for iedge in range(len(bin_edges)-1):
         temp = []
         for index,value in enumerate(input_grid):
             if bin_edges[iedge]<value<bin_edges[iedge+1]:
                 temp.append(input_value[index])
-        out[iedge] = np.average(temp)
-    return out
+        output_value[iedge] = np.average(temp)
+    return output_value
