+++ nemesispy/data/__init__.py
@@ -1,5 +1,0 @@
-#!/usr/local/bin/python3
-# -*- coding: utf-8 -*-
-"""
-Reference data files.
-"""
