### 2024-08-09 jingxuan97/nemesispy2022#6 @JBorrow
Can this (and the other related issues) now be closed?
