### 2024-08-09 jingxuan97/nemesispy2022#6 @Jingxuan97
Hi Josh,

Yes, I have addressed the comment and added the figure to the online
tutorial.

Best regards,
Jingxuan
