### 2024-08-09 jingxuan97/nemesispy2022#6 @JBorrow
Ok - please feel free to close them! It's not only good practice but will allow us to move forward with the review easier.
