### 2024-08-09 #6874 @JBorrow
Fantastic, thank you! @Jingxuan97, the next steps are as follows:

1. I will review the paper to make sure everything is ok. You can expect an automated message to come through letting us know if any DOIs are broken.
2. You should make a new release of the software and post a source deposit to Zenodo or Figshare.

See here: https://joss.readthedocs.io/en/latest/submitting.html#review-process.
