### 2024-08-14 #6874 @JBorrow
Hi @Jingxuan97 a reminder that we are awaiting a Zenodo deposit of the final version of the software with the following criteria:

- archive names (including order) and title in archive matches those specified in the paper
- archive uses the same license as the repo and is OSI approved as open source

We can then proceed to publication with some minor final checks.
