### 2024-09-03 #6874 @JBorrow
Hi @Jingxuan97, please make sure the deposit has the same name as the paper (i.e. NEMESISPY: A Python...), and that the authors names (not GitHub user names) are used. See e.g. https://zenodo.org/records/6447216 and https://joss.theoj.org/papers/10.21105/joss.04240
