### 2024-09-04 #6874 @JBorrow
@Jingxuan97 I'd recommend you set the version in your setup.py to be compatible with the zenodo deposit.
