### 2024-09-08 #6874 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
